clean up main.rs by moving "unit" procedural logic to rua_environment.rs
